% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNonExon.R
\name{countNonExon}
\alias{countNonExon}
\title{count the numbers, the position, percentage of introns for non-exon sequences on reference sequences.}
\usage{
countNonExon(findDataFrame, verifyDataFrame = data.frame())
}
\arguments{
\item{findDataFrame}{the data frame that store the information for funciton to analyze. Should compute from \code{findNonExon} or relateive functions.}

\item{verifyDataFrame}{the data frame that store the information for funciton to analyze. Should compute from \code{verifyNonExon} or relateive functions.}
}
\value{
A list of list with reference name, alignment position, alignment counts and percentage of introns if there is a verifyDataFrame.
}
\description{
\code{countNonExon} is a function that is used to count the number of non-exon sequences on specific position of reference sequences and output the percentage of introns in those non-exon sequences.
}
\details{
Details.
}
\section{usage}{
 countNonExon(findDataFrame).
}

\examples{
\dontrun{
dataFram <- findNonExon("reads.fasta", "transcripts.fasta")
countNonExon(dataFram)
verifydataFram <- verifyNonExon("reads.fasta", "introns.fasta")
countNonExon(dataFram, verifydataFram)
}
}
\seealso{
\code{\link{findNonExon}}, \code{\link{verifyNonExon}} compute the input for this function.
}
