% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exNonexonmap.R
\name{exNonexonmap}
\alias{exNonexonmap}
\title{An example of calling functions in my package.}
\usage{
exNonexonmap()
}
\value{
A hist of the position non-exon sequences on reference sequences.
}
\description{
\code{exNonexonmap} This is the example call of my functions with inputs from my extdata.
}
\examples{
exNonexonmap()
}
