% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNonExon.R
\name{findNonExon}
\alias{findNonExon}
\title{\code{findNonExon} is applied to find the non-exon sequences in reads.}
\usage{
findNonExon(readFile, referenceFile)
}
\arguments{
\item{readFile}{the reads file that is used as query, a string.}

\item{referenceFile}{the file usually transcripts that is used as reference, a string.}
}
\value{
a data frame.
}
\description{
\code{findNonExon} use \code{positionNonExon} to align the reads with transcripts and output a BAM file named "findNonExonTranscripts.BAM". The return value is a data frame with all position of alignment and number of matches and unmatches. \code{findNonExon} can be called by user directly.
}
\examples{
\dontrun{
dataFram <- findNonExon("reads.fasta", "transcripts.fasta")
}
}
\seealso{
\code{\link{usePosition}} the closure of \code{\link{positionNonExon}}
}
