% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainNonexonmap.R
\name{mainNonexonmap}
\alias{mainNonexonmap}
\title{An main function to plot the non-exon on reference gene.}
\usage{
mainNonexonmap(readsFile, referencesFile, intronsFile = "")
}
\arguments{
\item{readsFile}{The file that store the read sequences, a string.}

\item{referencesFile}{The file that store the reference sequences, a string.}

\item{intronsFile}{The file that contains the intron sequences, a string.}
}
\value{
A hist of the position non-exon sequences on reference sequences.
}
\description{
\code{mainNonexonmap} This is the main function for my functions in this package. The function will plot the non-exon position on reference gene and if they are introns (if intronFile exist).
}
\examples{
\dontrun{
readsFile <- system.file("extdata/testdata", "RRHreads.fasta", package = "nonexonmap")
transcriptsFile <- system.file("extdata/testdata", "RRHtranscript.fasta", package = "nonexonmap")
intronsFile <- system.file("extdata/testdata", "RRHintrons.fasta", package = "nonexonmap")
mainNonexonmap(readsFile, transcriptsFile)
mainNonexonmap(readsFile, transcriptsFile, intronsFile)
}
}
