% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonexonmap.maf}
\alias{nonexonmap.maf}
\title{NOSUCH NON-EXON MAP data.}
\format{Example A MAF formatted file with 15 rows and 32 columns:
\describe{
  \item{exampleHUGO_Symbol}{example this is a HGNC gene symbol}
  \item{exampleENTREZ_GENE_ID}{this is an ID}
  ...
}}
\source{
\url{https://github.com/VVVVVan/nonexonmap.git}
}
\description{
nonexonmap plots the map of non-exon sequences existed in reads in the final transcript sequences assembled by De Novo (Trinity). Since that the output file is just palin text, the plot and map of reads is an improvement. It analyzes the position of non-exon sequences on reference sequences. Furthermore, it checkes if the non-exon are introns and give the percentage of introns in the non-exon sequences. Rsubread and Rsamtools are used to align the sequences. A chart is produced for visualization.
}
\examples{
\dontrun{
system.file("extdata", "nonexonmap.maf", package="nonexonmap")
fPath <- system.file("extdata", "nonexonmap.maf", package="nonexonmap")
myExample <- read.delim(fPath, stringsAsFactors = FALSE)

}
}
