% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNonExon.R
\name{plotNonExon}
\alias{plotNonExon}
\title{plot the information of non-exon sequences on reference sequences.}
\usage{
plotNonExon(countLists)
}
\arguments{
\item{countLists}{a list of list with the non exon position and numbers to plot. Usually get from \code{countNonExon}.}
}
\value{
A list of hist data that could be used for test.
}
\description{
\code{plotNonExon} plot the numebr of non-exon sequences on the position of reference sequences. The percentage of the introns are also drawed, but dont't know how to combine these infomation on the plot at this point.
}
\examples{
\dontrun{
dataFram <- findNonExon("reads.fasta", "transcripts.fasta")
plotNonExon(countNonExon(dataFram))
verifydataFram <- verifyNonExon("reads.fasta", "introns.fasta")
plotNonExon(countNonExon(dataFram, verifydataFram))
}
}
\seealso{
\code{\link{countNonExon}} give the input of the function.
}
