% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNonExonApp.R
\name{runNonExonApp}
\alias{runNonExonApp}
\title{\code{runNonExonApp} launch the shiny app distributed with this package framework}
\usage{
runNonExonApp()
}
\value{
No return value but will pop up a shiny page.
}
\description{
\code{runNonExonApp} launches the shiny app for which the code has been placed in  \code{./inst/shiny-scripts/nonexonmapApp/}.
}
\examples{
\dontrun{
runNonExonApp()
}
}
