% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePosition.R
\name{usePosition}
\alias{usePosition}
\title{\code{usePosition} a closure of function \code{positionNonExon.R}.}
\usage{
usePosition(purpose)
}
\arguments{
\item{purpose}{output file form to determine the outputFile name, a string.}
}
\value{
a function.
}
\description{
\code{usePosition} form different output file for different purpose of using function \code{positionNonExon} To ensure that the output file of \code{positionNonExon} is in BAM format.
}
\examples{
\dontrun{
readsFile <- system.file("extdata/testdata", "RRHreads.fasta", package = "nonexonmap")
transcriptsFile <- system.file("extdata/testdata", "RRHtranscript.fasta", package = "nonexonmap")
dataFrame1 <- findNonExon(readsFile, transcriptsFile)
verifyNonExon <- usePosition("verify")
readsFile <- system.file("extdata/testdata", "RRHreads.fasta", package = "nonexonmap")
intronsFile <- system.file("extdata/testdata", "RRHintrons.fasta", package = "nonexonmap")
dataFrame2 <- verifyNonExon(readsFile, intronsFile)
}
}
\seealso{
\code{\link{positionNonExon}} find position of matches and unmatches.
}
