% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNonExon.R
\name{findNonExon}
\alias{findNonExon}
\title{Applied to find the non-exon sequences in reads.}
\usage{
findNonExon(readsFile, referencesFile)
}
\arguments{
\item{readsFile}{The file that store the read sequences, a string.}

\item{referencesFile}{The file that store the reference sequences usually transcripts, a string.}
}
\value{
A data frame contains the name of read, name of reference seqeunces, match/unmatch position and reference start position.
}
\description{
\code{findNonExon} use \code{positionNonExon} to align the reads with reference sequences and output a BAM file named "findNonExonTranscripts.BAM". The return value is a data frame with all position of alignment and number of matches and unmatches.
}
\examples{
\dontrun{
readsFile <- system.file("extdata/testdata", "RRHreads.fasta", package = "nonexonmap")
transcriptsFile <- system.file("extdata/testdata", "RRHtranscript.fasta", package = "nonexonmap")
dataFrame <- findNonExon(readsFile, transcriptsFile)
}
}
