% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainNonexonmap.R
\name{mainNonExonMap}
\alias{mainNonExonMap}
\title{An main function to plot the non-exon on reference gene.}
\usage{
mainNonExonMap(readsFile, transcriptsFile, intronsFile = "")
}
\arguments{
\item{readsFile}{The file that store the read sequences, a string.}

\item{transcriptsFile}{The file that store the reference sequences, a string.}

\item{intronsFile}{The file that contains the intron sequences, a string.}
}
\value{
A list of list with reference name, alignment position, alignment counts and percentage of introns if there is a verifyDataFrame.
}
\description{
\code{mainNonExonMap} This is the main function for my functions in this package. The function will plot the non-exon position on reference gene and if they are introns (if intronsFile exist).
}
\examples{
\dontrun{
readsFile <- system.file("extdata/testdata", "RRHreads.fasta", package = "nonexonmap")
transcriptsFile <- system.file("extdata/testdata", "RRHtranscript.fasta", package = "nonexonmap")
intronsFile <- system.file("extdata/testdata", "RRHintrons.fasta", package = "nonexonmap")
mainNonExonMap(readsFile, transcriptsFile)
mainNonExonMap(readsFile, transcriptsFile, intronsFile)
}
}
