% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionNonExon.R
\name{positionNonExon}
\alias{positionNonExon}
\title{output the positions of start alignment on references and number of nucleotides on reads that matches and unmatches to references.}
\usage{
positionNonExon(readFile, referenceFile, outputFile)
}
\arguments{
\item{readFile}{The file that store the read sequences, a string.}

\item{referenceFile}{The file that store the reference sequences, a string.}

\item{outputFile}{The name of output file in BAM format, a string.}
}
\value{
A data frame.
}
\description{
\code{positionNonExon} uses \code{buildindex} and \code{align} to form an alignment file. From the file, output the position of non-exon sequences exist on the references sequences and the match/unmatch numbers in the reads itself.
}
\examples{
\dontrun{
positionNonExon("reads.fasta", "transcript.fasta", "outputReadsTranscript.BAM")
positionNonExon("reads.fasta", "intron.fasta", "outputReadsIntron.BAM")
positionNonExon("exon.fasta", "wholeGene.fasta", "outputExonWholegene.BAM")
}
}
\seealso{
\code{\link[Rsubread]{buildindex}} Build an index for read mapping to perform.

\code{\link[Rsubread]{align}} Align DNA and RNA sequencing reads and report.
}
