% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUseAndApply.R
\name{usePosition}
\alias{usePosition}
\title{A closure for function \code{positionNonExon}.}
\usage{
usePosition(purpose)
}
\arguments{
\item{purpose}{output file form to determine the outputFile name, a string.}
}
\value{
a function.
}
\description{
\code{usePosition} form different output file for different purpose of using function \code{positionNonExon} To ensure that the output file of \code{positionNonExon} is in BAM format.
}
\examples{
\dontrun{
readsFile <- system.file("extdata/testdata", "RRHreads.fasta", package = "nonexonmap")
transcriptsFile <- system.file("extdata/testdata", "RRHtranscript.fasta", package = "nonexonmap")
findNonExon <- usePosition("find")
dataFrame1 <- findNonExon(readsFile, transcriptsFile)
}
}
