% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePosition.R
\name{usePosition}
\alias{usePosition}
\title{\code{usePosition} a closure of function \code{positionNonExon.R}.}
\usage{
usePosition(purpose)
}
\arguments{
\item{purpose}{output file form to determine the outputFile name, a string.}
}
\value{
a function.
}
\description{
\code{usePosition} form different output file for different purpose of using function \code{positionNonExon} To ensure that the output file of \code{positionNonExon} is in BAM format.
}
\examples{
\dontrun{
findNonExon <- usePosition("find")
dataFram1 <- findNonExon("reads.fasta", "transcript.fasta")
verifyNonExon <- usePosition("verify")
dataFram2 <- verifyNonExon("reads.fasta", "introns.fasta")
}
}
\seealso{
\code{\link{positionNonExon}} find position of matches and unmatches.
}
