% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyNonExon.R
\name{verifyNonExon}
\alias{verifyNonExon}
\title{\code{verifyNonExon} is applied to verify if the non-exon sequences are introns.}
\usage{
verifyNonExon(readFile, referenceFile)
}
\arguments{
\item{readFile}{the reads file that is used as query, a string.}

\item{referenceFile}{the file usually introns that is used as reference, a string.}
}
\value{
a data frame.
}
\description{
\code{verifyNonExon} use \code{positionNonExon} to align the reads with introns and output a BAM file named "verifyNonExonIntrons.BAM". The return value is a data frame with all position of alignment and number of matches and unmatches. \code{verifyNonExon} can be called by user directly.
}
\examples{
\dontrun{
dataFram <- verifyNonExon("reads.fasta", "introns.fasta")
}
}
\seealso{
\code{\link{usePosition}} the closure of \code{\link{positionNonExon}}
}
